+++ CHANGELOG.md
@@ -24,7 +24,9 @@
   pair (submitted on the command line as "`G:4:Image:12345`" or similar, passed on to
   the decorated function as `obj_type` and `obj_id`).
 * `hrm_omero.misc.printlog()` can be used to push a message to the log and stdout.
-* `hrm_omero.misc.parse_id_str()` validates the ID of a targeted OMERO object.
+* A class `hrm_omero.misc.OmeroId` provides parsing, validation and access to all
+  group-qualified OMERO object IDs (commonly passed as function parameter `id_str`
+  throughout the existing code).
 * The CLI has a new optional parameter `--dry-run` that prevents any action from being
   performed, instead the name of the function and the corresponding parameters that
   would be called are printed to stdout.